♻️ consolidation transaction update application
